requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.5.1
argparse==1.2.1
diff-match-patch==20110725.1
django-grappelli==2.5.3
django-import-export==0.2.3
django-redis-cache==0.13.0
hiredis==0.1.4
numpy==1.8.1
pandas==0.14.1
patsy==0.3.0
python-dateutil==2.2
pytz==2014.4
redis==2.10.1
scipy==0.14.0
six==1.7.3
sqlparse==0.1.11
statsmodels==0.5.0
tablib==0.10.0
wsgiref==0.1.2
